backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
sqlalchemy==2.0.35
alembic==1.13.0
psycopg2-binary==2.9.10
redis==5.0.1
celery==5.3.4
boto3==1.35.0
Pillow==11.0.0
numpy==1.26.4
pydantic==2.10.0
pydantic-settings==2.1.0
stripe==7.8.0
sendgrid==6.10.0
python-dotenv==1.0.0
email-validator==2.1.0
